\name{get_shark_path}
\alias{get_shark_path}
\title{
Extract the Shark output path from a Shark configuration file
}
\description{
Given a Shark configuration file, this function reads the relevant entries and
returns the root output directory used by Shark.
}
\usage{
get_shark_path(config_file)
}
\arguments{
  \item{config_file}{
Character scalar; path to a Shark configuration file.
}
}
\author{
Rodrigo Tobar
}
\examples{
# Nothing yet!
}
