\name{get_shark_path}
\alias{get_shark_path}
\title{
Extract the shark path out of a shark configuration file
}
\description{
Given a shark configuration file, this funcition reads the necessary configuration items
and returns the root directory of shark's outputs.
}
\usage{
get_shark_path(config_file)
}
\arguments{
  \item{config_file}{
Character scalar; path to a shark configuration file.
}
}
\author{
Rodrigo Tobar
}
\examples{
# Nothing yet!
}
